Feedback on defensive actions

+++ blog/views.py
@@ -89,36 +89,41 @@ def post_like(request, slug, *args, **kwargs):
     return HttpResponseRedirect(reverse('post_detail', args=[slug]))
 
 
-def comment_delete(request, slug, id, *args, **kwargs):
+def comment_delete(request, slug, comment_id, *args, **kwargs):
     """
     view to delete comment
     """
     queryset = Post.objects.filter(status=1)
     post = get_object_or_404(queryset)
-    comment = post.comments.filter(id=id)
-    comments = post.comments.filter(approved=True).order_by("-created_on")
-    comment.delete()
+    comment = post.comments.filter(id=comment_id).first()
 
-    messages.add_message(request, messages.SUCCESS, 'Comment deleted!')
+    if comment.name == request.user.username:
+        comment.delete()
+        messages.add_message(request, messages.SUCCESS, 'Comment deleted!')
+    else:
+        messages.add_message(request, messages.ERROR, 'You can only delete your own comments!')
 
     return HttpResponseRedirect(reverse('post_detail', args=[slug]))
 
 
-def comment_edit(request, slug, id, *args, **kwargs):
+def comment_edit(request, slug, comment_id, *args, **kwargs):
     """
     view to edit comments
     """
-    queryset = Post.objects.filter(status=1)
-    post = get_object_or_404(queryset, slug=slug)
-    comment = post.comments.filter(id=id).first()
+    if request.method == "POST":
 
-    comment_form = CommentForm(data=request.POST, instance=comment)
-    if comment_form.is_valid():
-        comment = comment_form.save(commit=False)
-        comment.post = post
-        comment.approved = False
-        comment.save()
+        queryset = Post.objects.filter(status=1)
+        post = get_object_or_404(queryset, slug=slug)
+        comment = post.comments.filter(id=comment_id).first()
 
-    messages.add_message(request, messages.SUCCESS, 'Comment Updated!')
+        comment_form = CommentForm(data=request.POST, instance=comment)
+        if comment_form.is_valid() and comment.name == request.user.username:
+            comment = comment_form.save(commit=False)
+            comment.post = post
+            comment.approved = False
+            comment.save()
+            messages.add_message(request, messages.SUCCESS, 'Comment Updated!')
+        else:
+            messages.add_message(request, messages.ERROR, 'Error updating comment!')
 
     return HttpResponseRedirect(reverse('post_detail', args=[slug]))
